Add mechanism for selecting events in the map
If you hover over an event graphic in the map, it'll now show a magenta
outline to show that it is selected (for things like moving it around,
editing the event or deleting it).

If there are multiple overlapping graphics, the one with the highest ID
will be selected because event graphics are rendered in order of ID.

Alternatively you can also hover over the actual tile on which the event
is located, which allows you to select an event that is completely
covered by another event's graphic without having to move the covering
graphic aside. A green border is rendered around the selected event's
tile to make this user-friendly.