Use a vector to store the events instead of a slab
A slab does not support inserting elements at arbitrary positions, which
we'll need to do if we want to change the ID of an event.

Also, we already know where all of the unused IDs are since we have to
iterate through all the events every time we draw the map, so a regular
vector will not lose any performance compared to a slab.