Enable WASM threading and disable WASM audio
I turned on some compiler flags to enable experimental support for
atomics and shared memory in WebAssembly, because this will probably be
needed if we end up using web workers to handle the filesystem
operations in WebAssembly builds.

Unfortunately, rodio doesn't support web workers, and in fact rodio
crashes Luminol in WebAssembly when I simply turn on those compiler
flags without making any changes to the code!

The audio in WebAssembly builds will have to be reimplemented without
rodio. I don't think this will be too hard, given that all we really use
rodio for is straightforward audio playback...