Refactor to not use `eframe::Frame`
I'm making a completely custom egui app runner for WebAssembly builds,
since eframe's web runner makes such widespread assumptions about the
availability of certain JavaScript APIs that are only present outside of
web workers that it's unsalvageable when it comes to getting it to run
in a web worker. (Both egui and wgpu can run in a web worker, though, so
a custom runner should work.)

But I ran into problems because implementing a custom runner requires
creating `eframe::Frame`s which can't be done outside of the eframe
crate.

In the future it would be best to stop using eframe entirely, but I
don't want to also have to rewrite the native runner right now so this
is the best I can do.