Rewrite web worker launcher code in Rust
The WebAssembly code now runs in both the main thread and the worker
thread with the same code and the same memory (i.e. multithreading),
but with different entry points.

The code on the main thread gets some information from the environment
and sends it to the code on the worker thread. The worker thread runs
the actual program.

The purpose of doing this is so that we can write event handlers in
Rust, for things like detecting user input on the canvas from the main
thread and sending it to the worker thread, or sending a request for a
filesystem operation from the worker thread to the main thread.