Add a hook to the main thread for detecting screen size changes
The main thread now listens for changes to the screen size and then
sends them to the worker thread so that it can resize the canvas to fill
the screen.

We use a `std::sync::mpsc::Sender` on the main thread to send the new
screen sizes to a `std::sync::mpsc::Receiver` on the worker thread. This
has the benefit of being more convenient than `Worker.postMessage()` and
also is probably faster because we don't have to deal with the
serialization overhead associated with `Worker.postMessage()`.

The Rust senders and receivers internally use WebAssembly's experimental
atomics and shared memory features since I enabled these things in the
compiler flags. They need Cross-Origin Isolation to be enabled to work,
so don't remove the coi-serviceworker.js shim.