Fix memory leak in `WebWorkerRunner::setup_render_hooks()`
At least I *think* this was a memory leak.

Anyways, the use of `CustomFrame` was somehow causing the rendering to
slow down slightly every frame. So that makes me think there was a
memory leak somewhere.

I couldn't figure out how to fix it properly so I just removed
`CustomFrame` since we don't need it anyways. But the most important
thing is that it runs at full speed now!