Tilemap now changes the projection matrix instead of the viewport rect
Before, the tilemap implemented panning and scaling by simply panning
and scaling the viewport rectangle that it rendered onto. This meant
if any part of the tilemap were offscreen, that part of the viewport
would be offscreen, too.

The WebGPU standard actually specifies that you're not allowed to have
any part of the viewport outside of the screen! (i.e. the intersection
of the viewport rectangle and the screen rectangle must be equal to the
viewport rectangle)

Consequently, most web browsers will not allow you to do this.

The version of wgpu that we currently use, version 0.16.3, allows you to
do this on native builds, but only because the developers forgot to
implement a check for this. It's actually undefined behaviour - some
drivers will allow this and some won't. It will be explicitly disallowed
in a future release of wgpu and already is in their trunk branch.

So, I made a minor change to the tilemap so that its viewport rectangle
always stays on screen. Now, the viewport rectangle is always the same
as the UI rectangle that egui allocates for the tilemap, and the
viewport projection matrix changes instead to handle panning and
scaling.

This should be the last change needed for the map editor to work in
WebAssembly, although it also needs to be applied to the event graphics
and the tilepicker, which is not done yet.