Tilemap event graphics now use projection matrix
The event graphics on the tilemap now render using the area allocated by
egui for the map editor as the viewport and handle scaling and panning
by modifying the projection matrix.

The event graphics in the hover tooltips do not do this, they just
render using the tooltip as the viewport because the tooltip doesn't
normally go offscreen. For some reason, egui sometimes draws the tooltip
offscreen for one frame before snapping it back onscreen, so a check is
added to disable rendering the tooltip graphic unless the graphic is
fully onscreen.