Fix "os error 22" when saving on Unix-like OSes
The Rust documentation says you need to enable the write flag to use
create or truncate when writing to files, so using the create flag
without also using the write flag is probably undefined behaviour.

It seems there are still more problems preventing saving from working,
but I assume this is in the scope of Astrabit-ST/Luminol#45 so I will
make no further changes.