Add CI checks for web builds
This adds a check for `wasm32-unknown-unknown` and a Trunk build to the
checks workflow and to check.sh.

The `-Z build-std=std,panic_abort` is required because the precompiled
standard library for this target doesn't ship with multithreading
support yet so we need to get Cargo to build it from source.

We don't need to do this for Trunk because I set up Trunk hooks to edit
.cargo/config.toml to enable these flags since Trunk doesn't support
passing custom flags to Cargo.