Fix map cursor moving when panning map using control key
This fixes an inconsistency in the map editor where panning the map by
dragging with the primary mouse button while the control key is held
down causes the map cursor to move if the mouse is hovering over an
event but does not cause the map cursor to move if the mouse is not
hovering over an event.

The map cursor now never moves when panning.

(cherry picked from commit 5f6f65ce3e213c3edd1b94a0e44ae4c663cd8a26)