feat(web): Implement touch events
This is something that eframe's original web runner supported that the
web worker runner didn't. It allows users to control Luminol via a
touchscreen display (like the one on the Microsoft Surface).

Chromium-based browsers have a "Device Mode" that emulates touch events
for those who lack devices with touchscreen displays.