Web refactor and custom eframe and egui-wgpu (#70)
* refactor: Add eframe and egui-wgpu 0.23.0 as subtrees

* docs: Add subtree merge instructions and egui licenses

* refactor: Use custom eframe and egui-wgpu

* refactor: Upgrade to wgpu 0.18

* refactor(web): Start moving the web runner into eframe

* refactor(web): Get rendering working again

Yes, I know it just renders a black rectangle right now. I'm getting to
the rest.

* refactor(web): Remove `luminol_app::CustomFrame`

* refactor(web): Implement mouse events

* perf(web): Replace eframe's time mutex with `AtomicF64`

Relaxed ordering should be fine here.

* chore(web): Remove unnecessary type annotations from closures

* refactor(web): Implement document events and window events

* fix(web): Use eframe's blurry canvas fix (round to even)

* refactor(web): Implement platform output and local storage

* chore(web): Rename `WebRunnerChannels` to `WorkerChannels` and remove `app_id`

* perf(web): Make `WORKER_DATA` a `Mutex` instead of `RwLock`

We never need to acquire a read lock on `WORKER_DATA`.

* feat(web): Implement touch events

This is something that eframe's original web runner supported that the
web worker runner didn't. It allows users to control Luminol via a
touchscreen display (like the one on the Microsoft Surface).

Chromium-based browsers have a "Device Mode" that emulates touch events
for those who lack devices with touchscreen displays.

* docs(web): Document public-facing parts of eframe to silence warnings

* chore(web): Silence unused variable warnings in eframe

* feat(web): Implement eframe's text agent

This text agent allows eframe to trigger the onscreen keyboard (for
devices that lack a keyboard) and/or input method editor (for e.g.
Chinese/Japanese/Korean input) when the user edits an egui `TextEdit` in
web builds.

* feat(web): Add screen reader support

This allows TTS to work properly in web builds if it's enabled.

* chore(web): Web runner channels are no longer optional

* refactor(web): Merge `channels` and `canvas` into one `state` variable

This reduces the amount of arguments we have to pass around.

* feat(web): Implement sending `wants_keyboard_input` to main thread

* fix(web): Intercept Ctrl+S, Ctrl+O and Ctrl+F

This prevents these keyboard shortcuts from performing their default
behaviour in web browsers so we can use them in Luminol.

* refactor(filesystem): Split web filesystem into smaller files

* refactor(filesystem): Add helper functions for sending commands

* chore(filesystem): Silence warnings in luminol-filesystem

* refactor(filesystem): Use async blocks to handle event errors

This encases all of the main thread event handlers for the web
filesystem in async blocks so that they can use the `?` operator for
proper error handling.

* style(filesystem): Remove unnecessary `return`

* fix: Enable `spin` feature of `flume`

Under rare circumstances, flume channel senders can block. This
absolutely cannot happen on the main thread, so we need to force flume
to use spinlocks instead of mutexes. Enabling spinlocks doesn't seem to
change the way flume receivers' blocking receive method works, which is
good.

* refactor(web): Add luminol- prefixes to eframe and egui-wgpu

* docs(web): Write what upstream commit eframe and egui-wgpu are based on

* refactor(flume): Add flume 0.11.0 as a subtree

* fix(flume): Fix flume's global lock in web builds

Flume's global spinlock uses `thread::sleep` which is not allowed in
WebAssembly, so I've removed `thread::sleep` from the spinlock.

* docs(flume): Add subtree merge instructions

* fix(flume): Disable clippy for flume

* fix: Fix failing eframe doctests

* chore(web): Remove `try_` from `MainState` inner borrows

Do or do not. There is no try.

Any instances of these borrows failing are unambiguous errors since they
all run on the main thread, so trying here is doing more harm than good.

* style(web): Remove unnecessary block

* refactor(flume): Move flume into a separate repo

* refactor(web): Move runner channel creation into eframe

This prevents us from having to change main.rs every time the channels
change.

* chore(filesystem): Fix project filesystem warnings in native builds

* refactor(filesystem): Move filesystem channel creation into luminol-filesystem

* chore: Remove `pub(self)`

I put these there originally so that I can do an easy replace with
`pub(super)` later if I needed to move things around.

* chore(web): Move web-sys features to the correct locations

* style(filesystem): Only the project folder name is now shown

This removes the name of the project file from the recent projects menu
in web for consistency with native.

* chore: Remove flume from rustfmt.toml

* chore: Add link to flume issue in Cargo.toml

---------

Co-authored-by: Lily Madeline Lyons <lily@nowaffles.com>