Unsaved changes modal (#74)
* fix: save empty audio file names as `''` instead of `nil`

* refactor: add a `modified` field to relevant parts of the data cache

* chore: switch to Git version of alox-48

* refactor: add `modified` field to maps as well

* perf: only save files that have been modified

System.rxdata is still saved every time to change the magic number.

* chore: bump alox-48 to 0.4.2

* fix: `push_to_history` actually pushes to history now

* style: use macros in the data cache to reduce boilerplate

* fix: fix borrow error when saving

* refactor: add `modified` field to `update_state`

This field keeps track of whether any file has been modified so that we
can show notifications when the user tries to close the window when
there are unsaved changes (not yet implemented).

* perf: only save System.rxdata if any other file was modified

* feat: modified maps now have "*" in the tab name

* refactor: add `modified` field to `rpg::System`

* feat: show unsaved changes dialogue in web builds on close

In web builds, this causes the browser to ask the user for confirmation
before closing Luminol if there are unsaved changes in the current
project.

* feat: show unsaved changes modal in native builds on close

* style: capitalize each word in "Unsaved Changes"

* style: rename `ProjectHandler` to `ProjectManager`

* style: don't show "Saving project..."

Because the filesystem is synchronous on every supported platform, this
message always appears at the same time as the "Saved project
successfully" message, so we don't really need to show both messages.

* refactor: move project picker code into `ProjectManager`

This allows us to deduplicate the code that opens the project picker and
also allows showing a confirmation modal if the user tries to open
another project while the current one has unsaved changes.

* refactor: move some project manager functions into `UpdateState`

* fix: fix wasm compile error

* refactor: move recent project loading code into `ProjectManager`

* refactor: move project closing code into `ProjectManager`

* style: fix warnings

* fix: close current project before loading a new one

* refactor: ask to save unsaved changes before creating new project

* refactor: remove portable-atomic from luminol-core

* style: merge `show_unsaved_changes_modal` and `handle_project_loading` into one function

* chore: fix clippy warnings

* feat: project manager now focuses and handles escape key

The unsaved changes modal now focuses the save button when it opens and
allows you to cancel by pressing the escape key on the keyboard. Also,
the menus in the top bar are now hidden when the modal opens.

* fix: project manager no longer has to wait 1 frame

If the user performs an action that changes the current project, it will
now be performed immediately (in the current frame) instead of being
performed at the beginning of the next frame.

* fix: prevent mouse interaction if project picker is open

* fix: handle saving errors properly in project manager

* fix: sucessfully -> successfully

* fix: reverse order of buttons in unsaved changes modal

* chore: fix another clippy warning

Sorry, I forgot to run clippy again after making those last few commits.