fix: remove eframe's blurry web canvas fix
My computer is one of those that get blurry web rendering when the width
of the canvas is an odd number of pixels. But in this modified version
of eframe's web runner, rounding the width to an even number of pixels
is actually what's causing the blurry rendering, so I've removed it.