refactor(filesystem): don't store in IndexedDB unless requested
In web builds, we usually don't need to store folder handles in
IndexedDB, so the old system of storing folder handles in IndexedDB by
default and deleting them from IndexedDB if requested was clunky.

This changes it so that folder handles are not stored by default. You
need to call `.save_to_idb()` to save it and get the key to restore it.