Update egui to 0.24 (#75)
* merge: update eframe and egui-wgpu from 0.23.0 to 0.24.1

* chore: update other dependencies

* chore: fix update-induced errors and warnings

* chore: rustfmt

* fix: remove eframe's blurry web canvas fix

My computer is one of those that get blurry web rendering when the width
of the canvas is an odd number of pixels. But in this modified version
of eframe's web runner, rounding the width to an even number of pixels
is actually what's causing the blurry rendering, so I've removed it.