Tilepicker scrollbar fix (#76)
* Fix tilepicker being too wide by default

In egui 0.24, the scrollbars no longer take up any space by default and
only appear when you hover over where the scrollbar is supposed to be.
So since Astrabit-ST/Luminol#75, the tilepicker width was incorrectly
including the width of the scrollbar in the width calculation, making
the tilepicker too wide. This removes the width of the scrollbar from
the tilepicker width formula so that the tilepicker default width and
maximum width are calculated correctly.

* Fix tilepicker scroll position changing when map is modified