fix(archive manager): don't persist collapsing header state
We don't want the state of the collapsing headers in the filesystem view
to be saved because we may be opening a different archive with the same
name, or we may be opening a modified version of the same archive. That
would lead to some errors when opening the view.