style(archive manager): future proofing and use generic IDs
This changes the filesystem view to use the indextree node IDs to
generate the egui IDs for the collapsing headers instead of using the
file paths to generate the egui IDs. This is so that we leave less junk
objects in the memory if the user opens a lot of different archives with
different names.

Also, the response tuple for the collapsing headers is now read using
pattern matching instead of by just indexing into the tuple, which makes
it easier to debug errors if egui decides to change the format of the
response tuple.