fix(filesystem): fix trie.rs `create_dir` edge case
I lied. THIS is the last edge case, for real this time!

This fixes a bug where the trie registers the root directory as a child
of itself if the user calls `create_dir` on the trie with any path and
then calls `create_dir` on the trie a second time with the empty path.