fix: properly set log verbosity to full in web builds
To set the verbosity in web builds, we need to use `.lib_verbosity`, not
`.verbosity`.

There will still be no backtrace displayed because we are using the
`panic_abort` panic handler instead of the normal `panic_unwind` handler
that native builds use. There is currently no support for `panic_unwind`
on the `wasm32-unknown-unknown` target unless we use `no_std`.