fix: terminal UI improvements
The terminal now immediately resizes to the correct size instead of
starting at the wrong size and taking several frames to resize.

The terminal size controls have been moved to the top of the terminal
window so that if the terminal is too big to fit on the screen, the user
can still access the controls to make it smaller. Also, the format is
now COLS x ROWS instead of ROWS x COLS.

The terminal also handles focus and hover correctly now. It won't change
the cursor unless the terminal is hovered, and it won't accept keyboard
input unless it's focused.

Also, the terminal no longer requests a rerender every frame. Now, it
only does that if the terminal's contents have been updated. For the
output window, this is done by making the tracing writer call
`ctx.request_repaint()` every time it writes to the log. For shells,
this is done by doing the same whenever there is at least one terminal
action received by the thread.