fix(web): fix environment being inconsistent between page reloads
This fixes a problem where sometimes, either cross-origin isolation
would not be enabled at all or sw.js would have its own COEP (but not
the COEP for other files) set to require-corp instead of credentialless.
This would be fixed by refreshing the page a number of times.

Now, COI should always be enabled and the COEP for sw.js should always
be credentialless, without the user having to refresh the page: the page
is automatically refreshed.