feat: truncate selectable label text instead of wrapping
This makes it so that the text in the selectable labels in the database
editors and the archive manager truncates instead of wrapping.

Before, it would wrap, meaning if the text was too long to fit in the
layout, it would wrap onto a new line of text. Now, the text will be
cut off with "..." if it's too long and will only occupy one line.

This makes the heights of the selectable labels uniform so that we can
use `.show_rows` on certain scroll areas instead of `.show`.