fix: shift selection now selects normally regardless of search
If you use shift selection in an `IdVecSelection` while searching, it
it should now select all entries in the range instead of only those in
the range that are visible in the search.

Someone might want to search for specific IDs as the start and end
endpoints of a large selection, so the new behaviour or selecting all
entries in the range would be more helpful than the old one.