perf: cache search results instead of recomputing every frame
This optimizes the search functionality in the combo boxes and in the
id_vec selection to cache the search results whenever the search query
or the items are changed and then use the cached results.

In order to track when items are changed, I've added a new field into
the `ModifiedState` that tracks if the data cache was modified during
the current frame.