perf: remove remaining O(n) database operations that run every frame
This commit optimizes the database editors further by removing a
linear-time check for if an id_vec is sorted that was running every
frame. Now, this check only runs when a database entry is modified.

The state editor was also cloning the names of every state every frame
to get around the borrow checker. The cloning of every state's name has
been replaced with an O(1) operation.