Implement some of the database editors (#99)
* refactor: separate out the code common to all database editors

This commit moves the generic skeleton code for all database editors
from the item editor into a new file,
crates/components/src/database_view.rs.

* refactor: move `Scope` and `Occasion` into crates/data/src/shared/mod.rs

* feat(skills): implement skill editor

* fix: fix ID collision in database view sidepane

* feat: outline where all the editors should go in the Data tab

* feat(weapons): implement weapon editor

* feat(armor): implement armor editor

* style(armor): rename "Armors" to "Armor"

* feat(states): implement state editor

* style(items,armor): increase consistency of naming

* style(states): move some widgets around

* feat(actors): start implementing the actor editor

* style(actors): replace drag values with sliders

* fix(actors): fix starting weapon/armor selection using wrong IDs

* fix(actors): only allow equippable weapons/armor to be selected

* fix(actors): mark project as modified when changing starting equipment

* feat: focus search box when `OptionalIdComboBox` opens

* feat(actors): implement drawing the parameter graphs

* style: add vertical spacing between fields

* fix(actors): anti-alias the left and right edges of the graphs

* fix(actors): fix off-by-one when calculating graph dimensions

* feat(classes): start implementing the class editor

* feat(classes): implement editing class skills

* feat(classes): implement element and state rank selection

* refactor: id_vec selection now always acts like control key is held

* fix(classes): improve consistency of invalid rank handling

* fix: fix off-by-one error in element selection logic

* fix: fix off-by-one error when displaying IDs

* refactor(classes): split out the collapsing editor code for the skills

* feat(enemies): start implementing the enemy editor

* feat(enemies): implement action editing for enemies

* style(enemies): reformat action fields a little

* fix(enemies): disable '(None)' options for treasure IDs

* style(enemies): flip stripe colors

* fix(enemies): display switch numbers in action conditions with 4 digits

* fix: increase minimum width of the database view

* refactor(classes,enemies): move collapsing view into separate functions

This commit moves the headers and bodies of the collapsing views for
these two editors into new functions to reduce the indentation levels in
the code.

* style(classes,enemies): add space under collapsing view delete button

* feat: truncate selectable label text instead of wrapping

This makes it so that the text in the selectable labels in the database
editors and the archive manager truncates instead of wrapping.

Before, it would wrap, meaning if the text was too long to fit in the
layout, it would wrap onto a new line of text. Now, the text will be
cut off with "..." if it's too long and will only occupy one line.

This makes the heights of the selectable labels uniform so that we can
use `.show_rows` on certain scroll areas instead of `.show`.

* feat: implement setting maximum number of database entries

* fix: use drag value height when calculating database view max height

* feat(actors): implement EXP curve viewer

* fix(actors): don't allow setting final level lower than initial level

* style: add margins on left and right of striped backgrounds

* style(actors,enemies): merge EXP curve rows and treasure rows

* style: change some margins in `OptionalIdVecComboBox`

* fix: change the lower bound for change maximum from 0 to 1

* fix: remove the bounds from the drag value for change maximum

We've reached a consensus that the upper bound of 999 is arbitary.

It would still be a good idea to limit this to prevent out-of-memory
crashes, but I don't know what to use as the upper bound.

* fix: keep ID intact when resetting database entries to default

* perf: put `IdVecSelection` into a scroll area

* fix(classes,enemies): resize rank tables when maximum number of states changes

* fix: fix `IdVecSelection` being broken when searching

* fix: shift selection now selects normally regardless of search

If you use shift selection in an `IdVecSelection` while searching, it
it should now select all entries in the range instead of only those in
the range that are visible in the search.

Someone might want to search for specific IDs as the start and end
endpoints of a large selection, so the new behaviour or selecting all
entries in the range would be more helpful than the old one.

* style: remove horizontal spacing from combo box stripes

This commit removes the margins from the stripes in the combo box lists
so that the items in the lists are the same width as the search box.

`ui.with_stripe()` is now split into `ui.with_padded_stripe()` which has
margins and `ui.with_stripe()` which doesn't.

* style: decrease space under search box in `IdVecSelection` slightly

* perf: cache search results instead of recomputing every frame

This optimizes the search functionality in the combo boxes and in the
id_vec selection to cache the search results whenever the search query
or the items are changed and then use the cached results.

In order to track when items are changed, I've added a new field into
the `ModifiedState` that tracks if the data cache was modified during
the current frame.

* feat: add warning if user tries to set maximum above 999

* perf: remove remaining O(n) database operations that run every frame

This commit optimizes the database editors further by removing a
linear-time check for if an id_vec is sorted that was running every
frame. Now, this check only runs when a database entry is modified.

The state editor was also cloning the names of every state every frame
to get around the borrow checker. The cloning of every state's name has
been replaced with an O(1) operation.

* feat: add search box to database view left pane

* fix: clear database view search box when view is closed

* fix: fix crash when decreasing maximum number of entries

* fix: fix database view left pane sometimes inverting stripe colors

* fix: justify the width of the above 999 warning