fix: clear browser cache when cached files are outdated
This changes the Trunk hooks to output a file called git-rev.txt into
the dist folder when compiling that contains the current git revision.

When loading a web build of Luminol, it will first try loading
git-rev.txt without caching to see what the latest git revision is. Then
if it doesn't match the cached git revision, it clears the cache and
reloads to get the latest version of Luminol.

The cache is also disabled if the git revision has "-modified" at the
end to ease development.

If there's an error fetching git-rev.txt, then the cache is not cleared.
This allows people to use their cached version of Luminol when they
don't have Internet access or when the website is down.