fix: fix edge case in `get_scripts_from_filesystem`
This fixes a bug where if the project's script path is set to something
beginning with "Game", the script manager searches for that in the root
directory of the project instead of the Data directory of the project.

The script manager now searches for the project's script path in the
Data directory regardless of what the script path is set to.