Script manager (#129)
* feat: implement script manager for RPG Maker XP

* feat: script manager now works for VX/VXA scripts as well

* fix: fix error when extracting every script at the same time

* feat: script manager now supports JSON/YAML/RON

* feat: add scripts conversion to script manager

* chore: clippy

* feat: allow selecting .rgssad archive as scripts file

In the "Extract" and "Convert" parts of the script manager, you can now
select a .rgssad archive. It'll look for a scripts file inside that
archive automatically.

* feat: read/write script order to _scripts.txt

* fix: fix edge case in `get_scripts_from_filesystem`

This fixes a bug where if the project's script path is set to something
beginning with "Game", the script manager searches for that in the root
directory of the project instead of the Data directory of the project.

The script manager now searches for the project's script path in the
Data directory regardless of what the script path is set to.