fix: fix 1 frame delay when grid shader viewport size changes
This fixes a bug where when the size of the map editor changes, the grid
shader's internally stored value for the size of the map editor's
viewport doesn't get updated until one frame later if the shader is
using push constants. This caused the thicknesses of the grid lines to
be inaccurate for a frame after resizing.