Fix animation cell with rotation and flip displaying incorrectly
If a cell needs to be rotated and flipped, the flip is supposed to be
applied first and then the rotation is to be applied afterwards.

The sprite shader currently applies the rotation first and then flip, so
I fixed this by reversing the direction of rotation passed to the shader
if flip is needed. It can be shown that in two dimensions, a flip that
goes through the origin followed by a rotation about the origin is the
same thing as doing the rotation first and then the flip if you reverse
the direction of rotation.