Improve handling of invalid frame numbers and cells
If the selected frame or selected cell in the animation editor becomes
invalid after the selected animation or frame number changes, it'll now
save what the frame index or selected cell index was and restore them if
they become valid again.

This is similar to the mechanism that text editors use when you use the
arrow keys to move between lines of a text document. If the line that
you move the cursor onto has fewer columns than the column that the
cursor was previously on, it saves the column and restores it if you
move the cursor onto a line with enough columns later.