Determine if file is MIDI using magic number instead of extension
There's no reason to be determining if an audio file is a MIDI file
based on the file extension when rodio determines the type of an audio
file based on its contents.

This also frees us from having to manually determine if an audio file is
MIDI when the audio data is passed as a slice.