Replace `frame.cell_data.xsize()` with `frame.len()`
The `.xsize()` of `frame.cell_data` doesn't always correspond to the
maximum cell number in a frame because of how the frame resizing
function defined in crates/ui/src/windows/animations/util.rs works.

So for all the code that was incorrectly assuming `.xsize()` was related
to the maximum cell number, I replaced the `.cell_data.xsize()` with a
new `.len()` method that more accurately determines the correct value.